Revisit routing and keycloak login component
