+++ src/auth/index.ts
@@ -1,7 +1,6 @@
 // Import API service modules
 import { AuthContext, AuthProvider } from "./AuthContext";
 import { ProtectedRoute } from "./ProtectedRoute";
-import { KeycloakLogin } from "./KeycloakLogin";
 import { KeycloakLogout } from "./KeycloakLogout";
 
-export { AuthProvider, AuthContext, ProtectedRoute, KeycloakLogin, KeycloakLogout };
+export { AuthProvider, AuthContext, ProtectedRoute, KeycloakLogout };
